Add Items and Searches sub-components for Cart component
